+++ packages/core/CHANGELOG.md
@@ -9,6 +9,7 @@ Versioning].
 ## [Unreleased]
 
 - Add option to ignore strings to `SingleGroupMangleExpression`.
+- Fix mangling query selectors in nested CSS selectors.
 
 ## [0.1.19] - 2021-04-30